Move losses temporairly to a new file
